x/integrity/types/message_create_hash_test.go
package types

import (
	"testing"

	"github.com/chen7gx/integrity/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateHash_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateHash
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateHash{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateHash{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
